+++ noodles-vcf/CHANGELOG.md
@@ -4,10 +4,16 @@
 
 ### Changed
 
+  * vcf/header/fmt: Add specialized META record serializer (#203).
+
+    This no longer quotes the `Number`, `Type`, and `Values` field values.
+
   * vcf/header/record/value/map/other: Add specialized PEDIGREE value parser.
 
     When the input is VCF 4.2, this allows the `Child` or `Derived` field to
     act as the record ID in the value collection.
+
+[#203]: https://github.com/zaeleus/noodles/issues/203
 
 ## 0.39.0 - 2023-09-14